requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0
tokenizers>=0.12.0
torch-geometric>=2.1.0
networkx>=2.8.0
jieba>=0.42.1
pypinyin>=0.47.0
nltk>=3.7.0
spacy>=3.4.0
opencv-python>=4.6.0
Pillow>=9.2.0
clip-by-openai>=1.0.0
tqdm>=4.64.0
data_utils>=2.4.0
pyyaml>=6.0.0
hydra-core>=1.2.0
wandb>=0.13.0
tensorboard>=2.10.0
requests>=2.28.0
beautifulsoup4>=4.11.0
emoji>=2.0.0